Changed electron-builder to create a zip (for simplicity)
This was due to the fact that addons would be installed deep within the user's app data folder so users would need to search for them. I want addons to be easily accessible.

One issue with this is that auto-updating functionality is not possible with a "zip" build. So I may need to find a way to port to the NSIS installer eventually.